requirements.txt
grpcio==1.60.1
grpcio-tools==1.60.1
paho-mqtt==2.0.0
protobuf==4.25.2
